- Login fertig gestellt mit ordentlicher Token-Validierung (Backend wie Frontend) -> es wird nach erfolgreichem Login ein Token (UUID-String) im WebSocketConnector.js gespeichert sowie im TokenManager im Backend der Token zusammen mit dem Usernamen in einer HashMap gespeichert und abgefragt bei weiteren Requests<br> - TestLogin angelegt unter dataManager/testData/testLogin.json -> dort ist der erste User "test" mit dem Passwort "Test1234" hinterlegt und kann auch für die Profilseite verwendet werden<br> - GetProfileWorkflow implementiert, sodass nach erfolgreichem Login der User unter "Profile" angezeigt wird - Login sowie Profilansicht wurden erfolgreich mit dem Testdatensatz getestet

+++ Backend/src/main/java/de/ideaWatcher/dataManager/pojos/User.java
@@ -21,9 +21,9 @@ public class User implements IUser {
     private String language;
     private String pictureUrl;
     private List<IIdea> createdIdeas;
-    private int numberCreatedIdeas;
+    private Double numberCreatedIdeas;
     private List<IIdea> followedIdeas;
-    private int numberFollowedIdeas;
+    private Double numberFollowedIdeas;
 
     @Override
     public String getUserName() {
@@ -126,12 +126,12 @@ public class User implements IUser {
     }
 
     @Override
-    public int getNumberCreatedIdeas() {
+    public Double getNumberCreatedIdeas() {
         return this.numberCreatedIdeas;
     }
 
     @Override
-    public void setNumberCreatedIdeas(int numberCreatedIdeas) {
+    public void setNumberCreatedIdeas(Double numberCreatedIdeas) {
         this.numberCreatedIdeas = numberCreatedIdeas;
     }
 
@@ -146,12 +146,12 @@ public class User implements IUser {
     }
 
     @Override
-    public int getNumberFollowedIdeas() {
+    public Double getNumberFollowedIdeas() {
         return this.numberFollowedIdeas;
     }
 
     @Override
-    public void setNumberFollowedIdeas(int numberFollowedIdeas) {
+    public void setNumberFollowedIdeas(Double numberFollowedIdeas) {
         this.numberFollowedIdeas = numberFollowedIdeas;
     }
 
@@ -169,8 +169,8 @@ public class User implements IUser {
         this.language = "";
         this.pictureUrl = "";
         this.createdIdeas = new ArrayList<>();
-        this.numberCreatedIdeas = 0;
-        this.numberFollowedIdeas = 0;
+        this.numberCreatedIdeas = 0.0;
+        this.numberFollowedIdeas = 0.0;
         this.followedIdeas = new ArrayList<>();
     }
 }
